requirements.txt
requests
typer
rich
streamlit
plotly
pandas
networkx
python-dotenv
pydantic
fastapi
uvicorn
psycopg2-binary
psycopg2
aiohttp
apscheduler
langchain-classic
langchain-core
langchain-openai
langchain-anthropic
langchain-nvidia-ai-endpoints